Folder/home page and File page refresh design
