+++ frontend/src/components/folders.js
@@ -24,12 +24,10 @@ class FolderList extends React.Component {
         return (
         <>
             {this.state.folders.map(folder => (
-                <a className="block border-b cursor-pointer" href={`/collection/folder/${folder.name}/${folder.folder_id}`}>
-                    <div className={folder.name == this.props.selectedName ? styles.selected : styles.default}>
-                        <div className="flex flex-row items-center space-x-2">
-                            <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M3 7v10a2 2 0 002 2h14a2 2 0 002-2V9a2 2 0 00-2-2h-6l-2-2H5a2 2 0 00-2 2z"></path></svg>
-                            <strong className="flex-grow font-normal">{folder.name}</strong>
-                        </div>
+                <a className="block cursor-pointer border p-4 border-gray-400 rounded-md shadow-md hover:border-cornblue-400 hover:bg-cornblue-200 hover:text-white" href={`/collection/folder/${folder.name}/${folder.folder_id}`}>
+                    <div className="flex flex-row items-center space-x-2">
+                        <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M3 7v10a2 2 0 002 2h14a2 2 0 002-2V9a2 2 0 00-2-2h-6l-2-2H5a2 2 0 00-2 2z"></path></svg>
+                        <strong className="flex-grow font-normal">{folder.name}</strong>
                     </div>
                 </a>
             ))}
