+++ frontend/src/pages/ChatsPage.js
@@ -294,21 +294,33 @@ function ChatsPage() {
                 <>
                 {selectedFriend !== null && (
                     <>
-                    {showFriendInfo ? (
-                        <div className="flex flex-col mt-2 w-1/3 bg-white rounded-tl-xl rounded-bl-xl border-l border-r border-gray-400 shadow-2xl p-2">
-                            <div className="h-16 bg-white flex flex-row justify-center items-center border-b">
-                                <button className="absolute right-0 pr-2" onClick={() => setShowFriendInfo(false)}>
-                                    <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"></path></svg>
-                                </button>
-                                <div className="flex-none">
-                                    <h1 className="font-semibold">{selectedFriend.firstname} {selectedFriend.lastname}</h1>
+                    {showFriendInfo ? (  
+                        <div className="fixed z-10 inset-0 overflow-y-auto">
+                            <div className="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
+                                <div className="fixed inset-0 transition-opacity" aria-hidden="true">
+                                    <div className="absolute inset-0 bg-gray-500 opacity-75"></div>
                                 </div>
-                            </div>
-                            <div className="mt-8 flex flex-col items-center space-y-6 justify-center">
-                                <img src={"http://" + process.env.REACT_APP_HOST_IP + ":3030" + selectedFriend.profile_pic} className="object-cover w-32 h-32 rounded-full" />
-                                <div className="flex flex-col items-center space-y-2">
-                                    <p className="font-bold text-lg">{selectedFriend.firstname} {selectedFriend.lastname}</p>
-                                    <p className="text-md">{selectedFriend.email}</p>
+                                <div className="absolute right-0 top-0 h-screen 2xl:w-1/4 w-96  text-cornblue-200 bg-white border-l border-r border-gray-400 overflow-y-auto">
+                                    <div className="h-16 bg-cornblue-400 flex flex-row justify-center items-center border-b">
+                                        <button className="absolute right-0 pr-2" onClick={() => setShowFriendInfo(false)}>
+                                            <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"></path></svg>
+                                        </button>
+                                        <div className="flex-none">
+                                            <h1 className="font-semibold">{selectedFriend.firstname} {selectedFriend.lastname}</h1>
+                                        </div>
+                                    </div>
+                                    <div className="px-6 mt-8 flex flex-col items-center space-y-6 justify-center text-black">
+                                        <img src={"http://" + process.env.REACT_APP_HOST_IP + ":3030" + selectedFriend.profile_pic} className="object-cover w-32 h-32 rounded-full" />
+                                        <p className="font-bold text-xl">{selectedFriend.firstname} {selectedFriend.lastname}</p>
+                                        <div className="w-full flex flex-col space-y-2">
+                                            <div className="flex flex-row justify-between border-b">
+                                                <p>Media</p>
+                                                <button>
+                                                    <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 5l7 7-7 7"></path></svg>
+                                                </button>
+                                            </div>
+                                        </div>
+                                    </div>
                                 </div>
                             </div>
                         </div>
